Version 3.27.22 (based on bleeding_edge revision r21712)
Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@21713 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/arguments.h
@@ -21,6 +21,9 @@ namespace internal {
 //   Object* Runtime_function(Arguments args) {
 //     ... use args[i] here ...
 //   }
+//
+// Note that length_ (whose value is in the integer range) is defined
+// as intptr_t to provide endian-neutrality on 64-bit archs.
 
 class Arguments BASE_EMBEDDED {
  public:
@@ -50,12 +53,12 @@ class Arguments BASE_EMBEDDED {
   }
 
   // Get the total number of arguments including the receiver.
-  int length() const { return length_; }
+  int length() const { return static_cast<int>(length_); }
 
   Object** arguments() { return arguments_; }
 
  private:
-  int length_;
+  intptr_t length_;
   Object** arguments_;
 };
